requirements.txt
# Core FastAPI and database
fastapi==0.92.0
uvicorn[standard]==0.20.0
sqlalchemy==1.4.23
psycopg2-binary==2.9.5
alembic==1.10.2

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
itsdangerous==2.1.2

# Validation and serialization
pydantic==1.10.5
email-validator==1.3.1

# HTTP client
httpx==0.23.3

# Template engine
jinja2==3.1.2

# Environment variables
python-dotenv==1.0.0

# AWS S3 compatible storage (R2)
boto3==1.26.137

# Additional dependencies
cryptography==39.0.2
certifi==2022.12.7

# JSON handling (using built-in instead of orjson)
# Removed: orjson (causes Rust compilation issues)

# Development dependencies (optional for production)
# pytest==7.2.2
# pytest-asyncio==0.21.0